Edit option of ManageProducts is fully Fucntional & Changed index.html of Admin to Default.aspx

+++ myAmazon-v1/AdminPanel/AddNewProduct.aspx.cs
@@ -24,7 +24,7 @@ namespace myAmazon_v1.AdminPanel
                 try
                 {
                     conn.Open();
-                    sqlCmd.ExecuteNonQuery();
+                    adapter.Fill(table);
                     conn.Close();
                 }
                 catch (Exception ex)
@@ -32,20 +32,66 @@ namespace myAmazon_v1.AdminPanel
                     id_log_product.Text += "Error opting Brand Names: " + ex.ToString();
                     conn.Close();
                 }
-                adapter.Fill(table);
                 id_category_name.DataSource = table;
                 id_category_name.DataTextField = "Name";
                 id_category_name.DataValueField = "id";
                 id_category_name.DataBind();
                 id_category_name.Items.Insert(0, new ListItem("--Select--", "NA"));
                 id_brand_name.Items.Insert(0, new ListItem("--Select--", "NA"));
+
+                if (Session["isEdit"] != null && Session["isEdit"].Equals("1"))
+                {
+                    Page.Title = "Edit Product";
+                    id_page_title.Text = "Edit Product";
+                    id_submit_product.Text = "Update";
+
+                    cmd = "SELECT p.Name, p.Price, p.[Desc], p.CategoryId, p.BrandId FROM Product AS p WHERE p.id=" + Session["ProductId"];
+                    sqlCmd = new SqlCommand(cmd, conn);
+                    SqlDataReader reader = null;
+                    try {
+                        conn.Open();
+                        reader = sqlCmd.ExecuteReader();
+                        reader.Read();
+                        id_product_name.Text = reader["Name"].ToString();
+                        id_product_price.Text = reader["Price"].ToString();
+                        id_category_name.SelectedValue = reader["CategoryId"].ToString();
+                        populateBrandDropDown();
+                        id_brand_name.SelectedValue = reader["BrandId"].ToString();
+
+                        //File handling for Description
+                        string path = reader["Desc"].ToString();
+                        StreamReader file = new StreamReader(Server.MapPath(path));
+                        id_product_desc.Text = file.ReadToEnd();
+                        file.Close();
+
+                        conn.Close();
+                    }
+                    catch (Exception ex)
+                    {
+                        id_log_product.Text = ex.ToString();
+                        conn.Close();
+                    }
+                }
+                
             }
         }
 
         protected void Press_Submit(object sender, EventArgs e)
         {
+            bool isEdit = false;
             SqlConnection conn = new SqlConnection(@"Data Source=DESKTOP-HO7NA1P;Initial Catalog=myAmazon;User ID=sa;Password=root");
-            SqlCommand insert = new SqlCommand("INSERT INTO Product(Name, [BrandId], [CategoryId], [Price]) OUTPUT inserted.id VALUES(@name, @brand, @category, @price)", conn);
+            SqlCommand insert = null;
+            string command;
+            if (Session["isEdit"] != null && Session["isEdit"].Equals("1"))
+                isEdit = true;
+
+            if (isEdit)
+                command = "UPDATE Product SET Name=@name, [BrandId]=@brand, [CategoryId]=@category, [Price]=@price WHERE id=" + Session["ProductId"];
+            else
+                command = "INSERT INTO Product(Name, [BrandId], [CategoryId], [Price]) OUTPUT inserted.id VALUES(@name, @brand, @category, @price)";
+
+            insert = new SqlCommand(command, conn);
+
             insert.Parameters.AddWithValue("@name", id_product_name.Text);
             insert.Parameters.AddWithValue("@brand", id_brand_name.Text);
             insert.Parameters.AddWithValue("@category", id_category_name.Text);
@@ -54,11 +100,19 @@ namespace myAmazon_v1.AdminPanel
             try
             {
                 conn.Open();
-                id = (int) insert.ExecuteScalar();
-                id_log_product.Text = "Successfully Inserted";
+                if (!isEdit)
+                {
+                    id = (int)insert.ExecuteScalar();
+                    id_log_product.Text = "Successfully Inserted";
+                    id_product_name.Text = "";
+                    id_product_price.Text = "";
+                }
+                else {
+                    id = Convert.ToInt32(Session["ProductId"]);
+                    insert.ExecuteNonQuery();
+                    id_log_product.Text = "Successfully Updated";
+                }
                 conn.Close();
-                id_product_name.Text = "";
-                id_product_price.Text = "";
             }
             catch (Exception ex)
             {
@@ -76,13 +130,23 @@ namespace myAmazon_v1.AdminPanel
                 {
                     id_log_product.Text += ex.ToString();
                 }
-                File.Create(Server.MapPath("~/ProductsData/" + id.ToString() + ".txt")).Close();
+                if(!File.Exists(Server.MapPath("~/ProductsData/" + id.ToString() + ".txt")))
+                    File.Create(Server.MapPath("~/ProductsData/" + id.ToString() + ".txt")).Close();
+                
                 File.WriteAllText(Server.MapPath("~/ProductsData/" + id.ToString() + ".txt"), id_product_desc.Text);
-                id_product_desc.Text = "";
-                conn.Open();
-                SqlCommand query = new SqlCommand("UPDATE Product SET [Desc] ='" + "~/ProductsData/" + id.ToString() + ".txt" + "' WHERE id=@cid", conn);
-                query.Parameters.AddWithValue("@cid", id);
-                query.ExecuteNonQuery();
+                if(!isEdit)
+                {
+                    id_product_desc.Text = "";
+                    conn.Open();
+                    SqlCommand query = new SqlCommand("UPDATE Product SET [Desc] ='" + "~/ProductsData/" + (isEdit ? Session["ProductId"] : id.ToString()) + ".txt" + "' WHERE id=@cid", conn);
+                    query.Parameters.AddWithValue("@cid", id);
+                    query.ExecuteNonQuery();
+                    conn.Close();
+                } else
+                {
+                    Session["isEdit"] = "0";
+                    Response.Redirect(@"..\AdminPanel\ManageProducts.aspx");
+                }
             }
             catch (Exception ex)
             {
@@ -92,6 +156,10 @@ namespace myAmazon_v1.AdminPanel
 
         protected void id_category_name_SelectedIndexChanged(object sender, EventArgs e)
         {
+            populateBrandDropDown();
+        }
+
+        private void populateBrandDropDown() {
             SqlConnection conn = new SqlConnection(@"Data Source=DESKTOP-HO7NA1P;Initial Catalog=myAmazon;User ID=sa;Password=root");
             string cmd = "SELECT id, Name FROM Brand WHERE CategoryId = @id";
             SqlCommand sqlCmd = new SqlCommand(cmd, conn);
@@ -104,7 +172,8 @@ namespace myAmazon_v1.AdminPanel
                 sqlCmd.ExecuteNonQuery();
                 conn.Close();
             }
-            catch (Exception ex) {
+            catch (Exception ex)
+            {
                 id_log_product.Text += "Error opting Brand Names: " + ex.ToString();
                 conn.Close();
             }
